move generate binding into a lib.rs for future reuse
